Streamlit_Chat/requirements.txt
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
streamlit==1.24.1
streamlit-chat==0.1.1
pandas==2.2.0
PyPDF2==3.0.1
python-dotenv==1.0.0
dataclasses-json==0.5.14